src/Mysql/requirements.txt
requests
pandas
fastapi
pydantic
uvicorn
mysql-connector-python